shared/models/feedback.py
from shared.db import db
from datetime import datetime
from sqlalchemy.orm import relationship

class Feedback(db.Model):
    __tablename__ = 'feedback'

    id = db.Column(db.Integer, primary_key=True)
    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
    content = db.Column(db.String(500), nullable=False)
    created_at = db.Column(db.DateTime, default=datetime.utcnow, nullable=False)
    user = relationship('User', backref='feedbacks')

    def to_dict(self):
        return {
            'id': self.id,
            'user_id': self.user_id,
            'content': self.content,
            'created_at': self.created_at.isoformat(),
            'user_name': self.user.username if self.user else None
        }

    def __repr__(self):
        return f'<Feedback {self.id} by {self.user_id}>'
