shared/requirements.txt
Flask
python-dotenv
Flask-SQLAlchemy
Werkzeug
